--- src/main/java/com/justifiedsolutions/openpdf/text/exceptions/InvalidPdfException.java
@@ -1,71 +1,0 @@
-/*
- * $Id: InvalidPdfException.java 3665 2009-01-26 22:32:15Z xlv $
- *
- * Copyright 2009 Bruno Lowagie
- *
- * The contents of this file are subject to the Mozilla Public License Version 1.1
- * (the "License"); you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.mozilla.org/MPL/
- *
- * Software distributed under the License is distributed on an "AS IS" basis,
- * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
- * for the specific language governing rights and limitations under the License.
- *
- * The Original Code is 'iText, a free JAVA-PDF library'.
- *
- * The Initial Developer of the Original Code is Bruno Lowagie. Portions created by
- * the Initial Developer are Copyright (C) 1999-2009 by Bruno Lowagie.
- * All Rights Reserved.
- * Co-Developer of the code is Paulo Soares. Portions created by the Co-Developer
- * are Copyright (C) 2000-2009 by Paulo Soares. All Rights Reserved.
- *
- * Contributor(s): all the names of the contributors are added in the source code
- * where applicable.
- *
- * Alternatively, the contents of this file may be used under the terms of the
- * LGPL license (the "GNU LIBRARY GENERAL PUBLIC LICENSE"), in which case the
- * provisions of LGPL are applicable instead of those above.  If you wish to
- * allow use of your version of this file only under the terms of the LGPL
- * License and not to allow others to use your version of this file under
- * the MPL, indicate your decision by deleting the provisions above and
- * replace them with the notice and other provisions required by the LGPL.
- * If you do not delete the provisions above, a recipient may use your version
- * of this file under either the MPL or the GNU LIBRARY GENERAL PUBLIC LICENSE.
- *
- * This library is free software; you can redistribute it and/or modify it
- * under the terms of the MPL as stated above or under the terms of the GNU
- * Library General Public License as published by the Free Software Foundation;
- * either version 2 of the License, or any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Library general Public License for more
- * details.
- *
- * If you didn't download this code from the following link, you should check if
- * you aren't using an obsolete version:
- * https://github.com/LibrePDF/OpenPDF
- */
-
-package com.justifiedsolutions.openpdf.text.exceptions;
-
-import java.io.IOException;
-
-/**
- * Typed exception used when opening an existing PDF document.
- * Gets thrown when the document isn't a valid PDF document.
- * @since 2.1.5
- */
-public class InvalidPdfException extends IOException {
-
-    /** a serial version UID */
-    private static final long serialVersionUID = -2319614911517026938L;
-
-    /**
-     * Creates an instance of a NoPdfException.
-     * @param    message    the reason why the document isn't a PDF document according to iText.
-     */
-    public InvalidPdfException(String message) {
-        super(message);
-    }
-}
