+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/IntHashtable.java
@@ -29,7 +29,7 @@
 
 package com.justifiedsolutions.openpdf.text.pdf;
 
-import com.justifiedsolutions.openpdf.text.error_messages.MessageLocalization;
+import com.justifiedsolutions.openpdf.text.MessageLocalization;
 
 import java.util.Arrays;
