+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/PRTokeniser.java
@@ -49,8 +49,8 @@
 
 package com.justifiedsolutions.openpdf.text.pdf;
 
-import com.justifiedsolutions.openpdf.text.error_messages.MessageLocalization;
-import com.justifiedsolutions.openpdf.text.exceptions.InvalidPdfException;
+import com.justifiedsolutions.openpdf.text.MessageLocalization;
+
 import java.io.IOException;
 /**
  *
