+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/PdfContentByte.java
@@ -49,24 +49,20 @@
 
 package com.justifiedsolutions.openpdf.text.pdf;
 
-import static com.justifiedsolutions.openpdf.text.pdf.ExtendedColor.MAX_COLOR_VALUE;
-import static com.justifiedsolutions.openpdf.text.pdf.ExtendedColor.MAX_FLOAT_COLOR_VALUE;
-import static com.justifiedsolutions.openpdf.text.pdf.ExtendedColor.MAX_INT_COLOR_VALUE;
-
+import com.justifiedsolutions.openpdf.text.MessageLocalization;
 import com.justifiedsolutions.openpdf.text.Rectangle;
-import com.justifiedsolutions.openpdf.text.error_messages.MessageLocalization;
-import com.justifiedsolutions.openpdf.text.exceptions.IllegalPdfSyntaxException;
-import com.justifiedsolutions.openpdf.text.pdf.internal.PdfXConformanceImp;
-import java.awt.Color;
+
+import java.awt.*;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import static com.justifiedsolutions.openpdf.text.pdf.ExtendedColor.*;
+
 /**
  * <CODE>PdfContentByte</CODE> is an object containing the user positioned
- * text and graphic contents of a page. It knows how to apply the proper
- * font encoding.
+ * text and graphic contents of a page. It knows how to apply the proper font encoding.
  */
 
 public class PdfContentByte {
@@ -389,7 +385,6 @@ public class PdfContentByte {
      * @param   blue    the intensity of blue. A value between 0 and 1
      */
     private void HelperRGB(float red, float green, float blue) {
-        PdfXConformanceImp.checkPDFXConformance(writer, PdfXConformanceImp.PDFXKEY_RGB, null);
         if (red < 0.0f)
             red = 0.0f;
         else if (red > MAX_FLOAT_COLOR_VALUE)
@@ -1033,7 +1028,6 @@ public class PdfContentByte {
      * @param color the color
      */
     public void setColorStroke(Color color) {
-        PdfXConformanceImp.checkPDFXConformance(writer, PdfXConformanceImp.PDFXKEY_COLOR, color);
         int type = ExtendedColor.getType(color);
         switch (type) {
             case ExtendedColor.TYPE_GRAY: {
@@ -1079,7 +1073,6 @@ public class PdfContentByte {
      * @param color the color
      */
     public void setColorFill(Color color) {
-        PdfXConformanceImp.checkPDFXConformance(writer, PdfXConformanceImp.PDFXKEY_COLOR, color);
         int type = ExtendedColor.getType(color);
         switch (type) {
             case ExtendedColor.TYPE_GRAY: {
@@ -1172,7 +1165,6 @@ public class PdfContentByte {
      * @param tint the tint if it is a spot color, ignored otherwise
      */
     void outputColorNumbers(Color color, float tint) {
-        PdfXConformanceImp.checkPDFXConformance(writer, PdfXConformanceImp.PDFXKEY_COLOR, color);
         int type = ExtendedColor.getType(color);
         switch (type) {
             case ExtendedColor.TYPE_RGB:
