+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/PdfImportedPage.java
@@ -49,7 +49,8 @@
 
 package com.justifiedsolutions.openpdf.text.pdf;
 
-import com.justifiedsolutions.openpdf.text.error_messages.MessageLocalization;
+import com.justifiedsolutions.openpdf.text.MessageLocalization;
+
 import java.io.IOException;
 
 /**
