+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/PdfPages.java
@@ -50,7 +50,8 @@
 package com.justifiedsolutions.openpdf.text.pdf;
 
 import com.justifiedsolutions.openpdf.text.ExceptionConverter;
-import com.justifiedsolutions.openpdf.text.error_messages.MessageLocalization;
+import com.justifiedsolutions.openpdf.text.MessageLocalization;
+
 import java.io.IOException;
 import java.util.ArrayList;
