+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/RandomAccessFileOrArray.java
@@ -50,17 +50,9 @@
 package com.justifiedsolutions.openpdf.text.pdf;
 
 import com.justifiedsolutions.openpdf.text.FontFactory;
-import com.justifiedsolutions.openpdf.text.error_messages.MessageLocalization;
-import java.io.ByteArrayOutputStream;
-import java.io.Closeable;
-import java.io.DataInput;
-import java.io.DataInputStream;
-import java.io.EOFException;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.RandomAccessFile;
+import com.justifiedsolutions.openpdf.text.MessageLocalization;
+
+import java.io.*;
 import java.net.URL;
 
 /**
