+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/TrueTypeFont.java
@@ -51,7 +51,7 @@ package com.justifiedsolutions.openpdf.text.pdf;
 
 import com.justifiedsolutions.openpdf.text.DocumentException;
 import com.justifiedsolutions.openpdf.text.ExceptionConverter;
-import com.justifiedsolutions.openpdf.text.error_messages.MessageLocalization;
+import com.justifiedsolutions.openpdf.text.MessageLocalization;
 
 import java.io.File;
 import java.io.IOException;
