+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/Type1Font.java
@@ -50,7 +50,7 @@
 package com.justifiedsolutions.openpdf.text.pdf;
 
 import com.justifiedsolutions.openpdf.text.DocumentException;
-import com.justifiedsolutions.openpdf.text.error_messages.MessageLocalization;
+import com.justifiedsolutions.openpdf.text.MessageLocalization;
 import com.justifiedsolutions.openpdf.text.pdf.fonts.FontsResourceAnchor;
 
 import java.io.ByteArrayOutputStream;
