+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/UnsupportedPdfException.java
@@ -47,14 +47,14 @@
  * https://github.com/LibrePDF/OpenPDF
  */
 
-package com.justifiedsolutions.openpdf.text.exceptions;
+package com.justifiedsolutions.openpdf.text.pdf;
+
 /**
- * Typed exception used when opening an existing PDF document.
- * Gets thrown when the document isn't a valid PDF document according to iText,
- * but it's different from the InvalidPdfException in the sense that it may
- * be an iText limitation (most of the times it isn't but you might have
- * bumped into something that has been added to the PDF specs, but that isn't
- * supported in iText yet).
+ * Typed exception used when opening an existing PDF document. Gets thrown when the document isn't a valid PDF document
+ * according to iText, but it's different from the InvalidPdfException in the sense that it may be an iText limitation
+ * (most of the times it isn't but you might have bumped into something that has been added to the PDF specs, but that
+ * isn't supported in iText yet).
+ *
  * @since 2.1.5
  */
 public class UnsupportedPdfException extends InvalidPdfException {
