+++ src/main/java/com/justifiedsolutions/openpdf/text/pdf/fonts/cmaps/CMap.java
@@ -30,7 +30,8 @@
  */
 package com.justifiedsolutions.openpdf.text.pdf.fonts.cmaps;
 
-import com.justifiedsolutions.openpdf.text.error_messages.MessageLocalization;
+import com.justifiedsolutions.openpdf.text.MessageLocalization;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
